- Added ByteSize class for allowing easier specification of sizes in config files, and used it for size configuration values. - SettingsDictionary now uses TypeConverter for type conversions.
git-svn-id: file:///mnt/c/Users/svgroot/Downloads/svn/repos/Jumbo/trunk@445 2f0f0f96-4b5b-11dd-99f4-89bc9257463c

+++ NameServer/NameServer.cs
@@ -57,7 +57,7 @@ namespace NameServerApplication
             _topology = new NetworkTopology(jumboConfig);
             _replicaPlacement = new ReplicaPlacement(Configuration, _topology);
             _replicationFactor = dfsConfig.NameServer.ReplicationFactor;
-            _blockSize = dfsConfig.NameServer.BlockSize;
+            _blockSize = (int)dfsConfig.NameServer.BlockSize;
             _fileSystem = new FileSystem(dfsConfig, replayLog);
             _fileSystem.FileDeleted += new EventHandler<FileDeletedEventArgs>(_fileSystem_FileDeleted);
             _fileSystem.GetBlocks(_blocks, _pendingBlocks);
